+++ app/routers/login.py
@@ -4,12 +4,8 @@ from fastapi import APIRouter, Depends, HTTPException, Request
 from fastapi.responses import RedirectResponse
 from sqlmodel import Session
 
-from app.internal.auth.auth import (
-    LoginTypeEnum,
-    RequiresLoginException,
-    auth_config,
-    get_authenticated_user,
-)
+from app.internal.auth.config import LoginTypeEnum, auth_config
+from app.internal.auth.login import RequiresLoginException, get_authenticated_user
 from app.util.db import get_session
 from app.util.templates import templates
 
@@ -21,19 +17,25 @@ async def login(
     request: Request,
     session: Annotated[Session, Depends(get_session)],
     error: Optional[str] = None,
+    redirect_uri: str = "/",
 ):
     login_type = auth_config.get(session, "login_type")
     if login_type != LoginTypeEnum.forms:
-        return RedirectResponse("/")
+        return RedirectResponse(redirect_uri)
 
     try:
         await get_authenticated_user()(request, session)
         # already logged in
-        return RedirectResponse("/")
+        return RedirectResponse(redirect_uri)
     except (HTTPException, RequiresLoginException):
         pass
 
     return templates.TemplateResponse(
         "login.html",
-        {"request": request, "hide_navbar": True, "error": error},
+        {
+            "request": request,
+            "hide_navbar": True,
+            "error": error,
+            "redirect_uri": redirect_uri,
+        },
     )
