+++ templates/login.html
@@ -19,6 +19,7 @@
     id="form"
     hx-target="#message"
   >
+    <input type="hidden" name="redirect_uri" value="{{ redirect_uri }}" />
     <h1 class="text-3xl">Login</h1>
     <label for="username">Username</label>
     <input
